requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
fastapi
uvicorn
alembic
pydantic
pydantic-settings
python-dotenv
pytest
